fix(user,workspace controller): issue of multiple imports in user and workspace services
