+++ .eslintrc
@@ -28,8 +28,8 @@
 		"@typescript-eslint/comma-dangle": "off", // Avoid conflict rule between Eslint and Prettier
 		"@typescript-eslint/consistent-type-imports": "error", // Ensure `import type` is used when it's necessary
 		"import/prefer-default-export": "off", // Named export is easier to refactor automatically
-		"simple-import-sort/imports": "error", // Import configuration for `eslint-plugin-simple-import-sort`
-		"simple-import-sort/exports": "error", // Export configuration for `eslint-plugin-simple-import-sort`
+		"simple-import-sort/imports": "off", // Import configuration for `eslint-plugin-simple-import-sort`
+		"simple-import-sort/exports": "off", // Export configuration for `eslint-plugin-simple-import-sort`
 		"@typescript-eslint/no-unused-vars": "off",
 		"unused-imports/no-unused-imports": "error",
 		// "unused-imports/no-unused-vars": ["error", { "argsIgnorePattern": "^_" }],
