+++ src/routes/api/v1/deploy.ts
@@ -12,31 +12,31 @@ dayjs.extend(localizedFormat);
 const controller = new DeployController();
 const router = express.Router();
 
-/**
- * Deploy from source code (git repository)
- */
-router.post("/", authenticate, async (req, res) => {
-	const { options } = req.body;
+router
+	/**
+	 * Deploy from source code (git repository)
+	 */
+	.post("/", authenticate, async (req, res) => {
+		const { options } = req.body;
 
-	// validation...
-	if (!options) return res.status(200).json({ status: 0, messages: [`Deploy "options" is required.`] });
-	if (!isJSON(options)) return res.status(200).json({ status: 0, messages: [`Deploy "options" is invalid (should be in JSON format).`] });
+		// validation...
+		if (!options) return res.status(200).json({ status: 0, messages: [`Deploy "options" is required.`] });
+		if (!isJSON(options)) return res.status(200).json({ status: 0, messages: [`Deploy "options" is invalid (should be in JSON format).`] });
 
-	// convert JSON to Object
-	req.body.options = JSON.parse(options);
+		// convert JSON to Object
+		req.body.options = JSON.parse(options);
 
-	// no need to wait :)
-	controller.deployFromSource(req.body).then((result) => log(`Finished deploying from source code:`, { result }));
+		// no need to wait :)
+		controller.deployFromSource(req.body).then((result) => log(`Finished deploying from source code:`, { result }));
 
-	res.status(200).json({ status: 1 });
-});
-
-/**
- * Deploy from image URL
- */
-router.post("/from-image", authenticate, async (req, res) => {
-	const result = await controller.deployFromImage(req.body);
-	res.status(200).json(result);
-});
+		res.status(200).json({ status: 1 });
+	})
+	/**
+	 * Deploy from image URL
+	 */
+	.post("/from-image", authenticate, async (req, res) => {
+		const result = await controller.deployFromImage(req.body);
+		res.status(200).json(result);
+	});
 
 export default router;
